iac/aws_batch/requirements.txt
aws-cdk-lib==2.160.0
constructs>=10.0.0,<11.0.0
boto3<1.34.132,>=1.34.70
botocore<1.34.132,>=1.34.70